\name{neftest-package}
\alias{neftest-package}
\alias{neftest}
\docType{package}
\title{
\packageTitle{neftest}
}
\description{
\packageDescription{neftest}
}
\details{
\tabular{ll}{
Package: \tab neftest\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2021-07-18\cr
License: \tab GPL (>= 2)\cr
}
}
\references{Authors (2021). Goodness of fit tests based on zero regression characterizations of Tweedie, Bar-Lev and Enis class of distributions. Manuscript.
}

\keyword{ package }
